requirements.txt
django
dj-rest-auth
django-allauth
djangorestframework
djangorestframework-simplejwt
Pillow